debug: make allThreadStop reason generic
When we receive a stopped event from a debug adapter, a
`allThreadStopped` flag can be set. When doing so, it feels weird to
display the reason of the thread that caused it on other threads.

This commit reports "PAUSED" for all the paused threads, but keeps
displaying the reason for the thread that caused the event.

Signed-off-by: Paul Maréchal <paul.marechal@ericsson.com>